Time: 0 ms (100.00%) | Memory: 25.3 MB (100.00%) - LeetSync

+++ 3471-minimum-average-of-smallest-and-largest-elements/minimum-average-of-smallest-and-largest-elements.cpp
@@ -2,7 +2,7 @@ class Solution {
 public:
     double minimumAverage(vector<int>& nums) {
         sort(nums.begin(),nums.end());
-        double v=DBL_MAX;
+        double v=INT_MAX;
         for(int i=0;i<nums.size()/2;i++)
         {
             if((nums[i]+nums[nums.size()-i-1])/2.0<v)
